requirements.txt
flask
flask-socketio
google-api-python-client
requests
python-dotenv
yt-dlp